pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-fragility"
version = "0.1.0"
description = "Visualizing AI fragility: adversarial attacks and assurance dashboard"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "AI Assurance Lab" }]
license = { file = "LICENSE" }
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "lightning>=2.2.0",
    "hydra-core>=1.3.2",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "polars>=0.20.0",
    "duckdb>=0.10.0",
    "wandb>=0.17.0",
    "optuna>=3.6.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.6.0",
    "pillow>=10.2.0",
    "torchmetrics>=1.3.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "pyarrow>=15.0.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = ["ruff>=0.6.0", "pytest>=8.0.0", "mypy>=1.10.0", "pytest-cov>=5.0.0"]

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["ai_fragility"]

[tool.hatch.build.targets.editable]
dev-mode-dirs = ["src"]
packages = ["ai_fragility"]

[tool.ruff]
target-version = "py311"
line-length = 100
extend-select = ["I"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "W",
    "C90",
    "N",
    "UP",
    "S",
    "A",
    "COM",
    "RET",
    "PL",
    "RUF",
]
ignore = ["D203", "D212"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["ai_fragility"]
